#537 WIP checkpoint
