+++ allensdk/test/brain_observatory/ecephys/stimulus_table/test_naming_utilities.py
@@ -1,0 +1,7 @@
+import pytest
+import pandas as pd
+
+from allensdk.brain_observatory.ecephys.stimulus_table.naming_utilities as nu
+
+
+def test_
